project/requirements.txt
Pillow==9.0.1
imagehash==4.2.1
